Move sales `showRowsAjax` to Reports' Module Sales Controller
