+++ app/Http/Controllers/SalesModule/SalesController.php
@@ -12,38 +12,6 @@ use Ramsey\Uuid\Uuid;
 
 class SalesController extends Controller
 {
-    public function showRowsAjax()
-    {
-        // Convert this SQL query below into a query builder
-        // SELECT invoices.*, customers.name as customer_name, SUM(total_price) as grand_total, COUNT(p.invoice_id) as no_products
-        // FROM `invoices`
-        // LEFT JOIN `customers` ON invoices.customer_id = customers.id
-        // LEFT JOIN (
-        //     SELECT (sales.quantity_sold * consigned_products.sale_price) as total_price, sales.quantity_sold, consigned_products.sale_price, invoice_id, sales.id as sales_id
-        //     FROM `sales`
-        //     LEFT JOIN consigned_products ON consigned_products.id = sales.consigned_product_id
-        // ) AS p ON p.invoice_id = invoices.id
-        // GROUP BY invoice_id;
-        $invoices = DB::table('invoices')
-            ->select(
-                'invoices.*',
-                'customers.name as customer_name',
-                DB::raw('SUM(total_price) as grand_total'),
-                DB::raw('COUNT(p.invoice_id) as no_products')
-            )
-            ->leftJoin('customers', 'invoices.customer_id', '=', 'customers.id')
-            ->leftJoin(
-                DB::raw('(SELECT (sales.quantity_sold * consigned_products.sale_price) as total_price, sales.quantity_sold, consigned_products.sale_price, invoice_id, sales.id as sales_id FROM `sales` LEFT JOIN consigned_products ON consigned_products.id = sales.consigned_product_id) AS p'),
-                'p.invoice_id',
-                '=',
-                'invoices.id'
-            )
-            ->groupBy('invoice_id')
-            ->get();
-
-        return $invoices;
-    }
-
     public function create()
     {
         return view('sales.create');
